requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
psycopg2-binary==2.9.7
python-jenkins==1.8.0
jira==3.5.0
APScheduler==3.10.4
python-dotenv==1.0.0
celery==5.3.1
redis==4.6.0
gunicorn==21.2.0